refactor(inference): split services/inference.py into focused modules (#554)
inference.py was 775 LOC mixing spaCy lifecycle, NER, RSS metadata
candidate construction, DB reads for recurring host + feed_speaker_cache,
confidence reconciliation, heuristic classification, speaker-slot
remapping, and DB writes. Splitting those concerns into dedicated
modules without changing behavior.

New layout under apps/pipeline/app/services/:

  inference_types.py     - CandidateName, InferenceResult, METADATA_SOURCES
  inference_ner.py       - extract_candidates (spaCy), _looks_like_person_name,
                           _podcast_person_to_candidate, extract_metadata_candidates
  inference_classify.py  - merge_candidates, classify_candidates
  inference_db.py        - get_recurring_host_name, get_feed_speaker_cache_priors,
                           write_speaker_names, plus their private min-count constants
  inference.py           - thin orchestrator: spaCy load/unload (keeps the
                           sys.modules[_nlp] reach), assign_speaker_slots,
                           and re-exports of the public surface

No behavior changes. No signature changes. All existing imports
(from app.services.inference import ...) continue to resolve via
re-export from inference.py. Docstrings explaining the MEDIUM vs HIGH
self-reinforcement guards (see PRD-04 A1, B1/B2/B3, C1/C2 and issue #530
reconciliation) moved with the code.

Safety net:
- tests/unit/test_inference.py (115 tests, 1372 LOC) passes untouched.
- Full pipeline unit suite: 698 passed, coverage 86.62% (gate 82%).

Per-module coverage under unit tests:
  inference_types     100%
  inference_ner        99%
  inference_classify   93%
  inference_db         92%
  inference              57% (spaCy load/unload only - integration path)
